backend/requirements.txt
fastapi
crewai
openai
python-dotenv
uvicorn
python-multipart
crewai[tools]
langchain-openai

